+++ src/main/kotlin/ua/com/pimenov/latte/data/LatteFileExtensions.kt
@@ -1,0 +1,3 @@
+package ua.com.pimenov.latte.data
+
+val TEST_FILE_EXTENSIONS = setOf("test.js", "test.jsx", "test.ts", "test.tsx", "spec.js", "spec.jsx", "spec.ts", "spec.tsx")